requirements.txt
streamlit
pandas
supabase
groq
python-dotenv
plotly
